back end, web and mobile

+++ server/src/routes/memories.ts
@@ -2,9 +2,20 @@ import { FastifyInstance } from 'fastify'
 import { z } from 'zod'
 import { prisma } from '../lib/prisma'
 
+
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+
+  app.addHook('preHandler',async (request) => {
+    await request.jwtVerify()
+  })
+
+  app.get('/memories', async (request) => {
+
+
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: 'asc',
       },
@@ -14,32 +25,36 @@ export async function memoriesRoutes(app: FastifyInstance) {
       return {
         id: memory.id, 
         coverUrl: memory.coverUrl,
-        excerpt: memory.content.substring(0, 115).concat('...'),
+        excerpt: memory.content.substring(0, 115).concat('...'), //criado para não mostrar o conteudo todo e quebrar com ... no final
       }
     })
   })
 
-  app.get('/memories/:id', async (request) => {
-    const paramsSchema = z.object({
-      id: z.string().uuid(),
+  app.get('/memories/:id', async (request,reply) => {
+    const paramsSchema = z.object({ //usando o zod pra validar o parametro
+      id: z.string().uuid(), //verifica se o id é uma string e uuid
     })
 
     const { id } = paramsSchema.parse(request.params)
 
-    const memory = await prisma.memory.findUniqueOrThrow({
+    const memory = await prisma.memory.findUniqueOrThrow({ //encontra a memoria ou joga erro na tela
       where: {
         id,
       },
     })
+
+    if(!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
     return memory
   })
 
   app.post('/memories', async (request) => {
     const bodySchema = z.object({
-      content: z.string(),
+      content: z.string(), 
       coverUrl: z.string(),
-      isPublic: z.coerce.boolean().default(false),
+      isPublic: z.coerce.boolean().default(false), //coerce converte o boolean
     })
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
@@ -49,14 +64,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: '5ac8562f-5c67-49da-8319-5646cf8b8b6d',
+        userId: request.user.sub,
       },
     })
 
     return memory
   })
 
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request,reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -71,7 +86,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findFirstOrThrow({
+      where: {
+        id,
+      }
+    })
+
+    if(memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
+     memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -85,12 +110,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory
   })
 
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request,reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
 
     const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findFirstOrThrow({
+      where: {
+        id,
+      }
+    })
+
+    if(memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
     await prisma.memory.delete({
       where: {
